Lift photo manage

+++ backend/controllers/LiftController.php
@@ -14,6 +14,8 @@ use yii\web\Response;
 
 class LiftController extends Controller
 {
+
+    public $imagesOnCreate = '';
     public function behaviors()
     {
         return [
@@ -33,6 +35,7 @@ class LiftController extends Controller
     }
     public function actionIndex()
     {
+       
         $lifts = Lift::find()->all();
         $types = [];
         $producers = [];
@@ -65,29 +68,6 @@ class LiftController extends Controller
             Yii::$app->session->setFlash('error', 'Немає виробників ліфтів');
             return  $this->redirect(['lift/index']);
         }
-        $model = new LiftForm;
-        if($model->load(Yii::$app->request->post()))
-        {
-            $model->imageFile = UploadedFile::getInstances($model,'imageFile');
-           if($imagePath=$model->upload())
-           {
-            $lift = new Lift;
-            $lift->description = $model->description;
-            $lift->type_id = $model->type_id ;
-            $lift->producer_id = $model->producer_id ;
-            $lift->url_image = json_encode($imagePath);
-            if($lift->save())
-            {
-            // Yii::$app->session->setFlash('success', 'Товар збережено в БД ');
-            }
-            else{
-                Yii::$app->session->setFlash('error', 'Помилка НЕ збережено в БД ');
-               }
-           }
-         return  $this->redirect(['lift/index']);
-          
-        }
-        
         foreach ($types as $type) {
             $type_array[$type->id] = $type->name;
         }
@@ -95,6 +75,43 @@ class LiftController extends Controller
         foreach ($providers as $provider) {
             $provider_array[$provider->id] = $provider->name;
         }
+        $model = new LiftForm;
+        if($model->load(Yii::$app->request->post()))
+        {
+            $model->imageFile = UploadedFile::getInstances($model,'imageFile');
+            if($imagePath=$model->upload())
+           {
+            $lift = new Lift;
+            $lift->description = $model->description;
+            $lift->type_id = $model->type_id ;
+            $lift->producer_id = $model->producer_id ;
+
+            $lift->url_image = json_encode($imagePath);
+            if($lift->save())
+            {
+                return  $this->redirect(['lift/'.$lift->id.'/update']);
+            // Yii::$app->session->setFlash('success', 'Товар збережено в БД ');
+            }
+            else{
+                Yii::$app->session->setFlash('error', 'Помилка НЕ збережено в БД ');
+               }
+           } else{
+           return $this->render('create', [
+            'model' => $model,
+            'types' => $type_array,
+            'producers' => $provider_array,
+            'initialPreview' => [],
+            'initialConfig' => [],
+            'lift_id'=> '',
+            'msg'=>'обов\'язкове поле',
+            'dangerStyle' =>'border: 1px solid red; border-radius: 10px;',
+        ]);
+    }
+   
+          
+        }
+        
+       
 
         return $this->render('create', [
             'model' => $model,
@@ -103,35 +120,26 @@ class LiftController extends Controller
             'initialPreview' => [],
             'initialConfig' => [],
             'lift_id'=> '',
+            'msg'=>'',
+            'dangerStyle' =>'',
         ]);
     }
     public function actionUpdate($id){
         $model = new LiftForm;
         $lift = Lift::findOne(['id' => $id]);
         if($model->load(Yii::$app->request->post()))
-        {
-            $model->imageFile = UploadedFile::getInstances($model,'imageFile');
-            $imagePath=$model->upload();
-           if($imagePath !== false)
-           {
+        {       
             $lift->description = $model->description;
             $lift->type_id = $model->type_id ;
-            $lift->producer_id = $model->producer_id ;
-            if($imagePath)
-            {
-                $image = json_decode($lift->url_image,true);
-                $imagePath = array_merge($image, $imagePath);
-                $lift->url_image = json_encode($imagePath);
-            }
-            
+            $lift->producer_id = $model->producer_id ;        
             if($lift->save())
             {
              //Yii::$app->session->setFlash('success', 'Товар збережено в БД ');
             }
             else{
                 Yii::$app->session->setFlash('error', 'Помилка НЕ збережено в БД ');
-               }
-           }
+            }
+           
          return  $this->redirect(['lift/index']);
           
         }
@@ -158,7 +166,7 @@ class LiftController extends Controller
             ];
         }
 
-        return $this->render('create', [
+        return $this->render('update', [
             'model' => $model,
             'producers' => $provider_array,
             'types' => $type_array,
@@ -191,22 +199,26 @@ class LiftController extends Controller
     public function actionFileDeleteLift($id){
         
         Yii::$app->response->format = \yii\web\Response::FORMAT_JSON;
-        var_dump("1");
+       // var_dump("1");
         if(isset($_POST['key'])){
             $image = $_POST['key'];
-            var_dump("2");
-            var_dump($_POST['key']);
-            unlink($_POST['key']);
+           // var_dump("2");
+           // var_dump($_POST['key']);
+           
+            
             $lift = Lift::findOne(['id' => $id]);
             $images = json_decode($lift->url_image,true);
-            $result = [];
-            foreach ($images as $value) {
-                if($image != $value){
-                    $result[] = $value;
+            if(count($images) > 1){
+                unlink($_POST['key']);
+                $result = [];
+                foreach ($images as $value) {
+                    if($image != $value){
+                        $result[] = $value;
+                    }
                 }
+                $lift->url_image = json_encode($result);
+                $lift->save();
             }
-            $lift->url_image = json_encode($result);
-            $lift->save();
         }
         return true;
     }
@@ -263,21 +275,21 @@ class LiftController extends Controller
         ]);
 
     }
-    public function actionUpload()
+    public function actionUpload($id)
     {
         if (isset($_POST)) {
-            $files= LiftController::SaveTempAttachments($_FILES);
+            $files= LiftController::SaveTempAttachments($_FILES, $id);
             $result = ['files'=>$files];
             Yii::$app->response->format = trim(Response::FORMAT_JSON);
-            return $result;
+            return  $result;
         }
 
     }
-    public static function SaveTempAttachments($attachments)
-    {
+    public static function SaveTempAttachments($attachments, $id)
+    {    $lift = Lift::findOne(['id' => $id]);
         $files[]='';
         $allwoedFiles=['jpg'];
-    
+        $imagePath=[];
         if (!empty($attachments)) {
             if (count($attachments['files']['name']) > 0) {
                 //Loop through each file
@@ -295,8 +307,11 @@ class LiftController extends Controller
                             //save the url and the file
                             $newFileName = Yii::$app->security->generateRandomString(40) . "." . $ext;
                             //Upload the file into the temp dir
-                            if (move_uploaded_file($tmpFilePath,  '../../images/lift/' . $newFileName)) {
+
+                          array_push($imagePath , '../../images/lift/' . $newFileName);
+                            if (move_uploaded_file( $tmpFilePath ,  '../../images/lift/' . $newFileName)) {
                                 $files[] =['fileName'=>$newFileName,'type'=>$ext,'size'=>(($size/1000)),'originalName'=>$shortname];
+                               
                             }
     
                         }
@@ -305,7 +320,20 @@ class LiftController extends Controller
             }
     
         }
-        return $files;
+        $image = json_decode($lift->url_image,true);
+        $imagePath = array_merge($image, $imagePath);
+        $lift->url_image = json_encode($imagePath);
+        if($lift->save())
+            {
+             //Yii::$app->session->setFlash('success', 'Товар збережено в БД ');
+            }
+            else{
+                Yii::$app->session->setFlash('error', 'Помилка НЕ збережено в БД ');
+               }
+           
+        $images = json_decode($lift->url_image,true);
+        return $images;
     }
+   
    
 }
